+++ pkg/analysis_server_plugin/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Improve performance of plugin analysis in cases where all of the enabled
   plugin rules in a given plugin isolate are "parse-only" rules which have a
   `true` value for `canUseParsedResult`.
+- Improve performance of plugin analysis in cases where plugins are enabled
+  in subdirectories of a workspace.
 - Require version `14.4.0-dev` of the `analyzer` package.
 - Require version `0.14.17-dev` of the `analyzer_plugin` package.